Moving/renaming some files & adding some safety guards to prevent game crashing
