+++ Renegade X Launcher/Converters/BitmapToImageSourceConverter.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Drawing;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+using System.Windows.Interop;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+
+namespace LauncherTwo
+{
+    // From http://stackoverflow.com/questions/5824706/how-to-bind-images-from-properties-resources-in-xaml
+    [ValueConversion(typeof(Bitmap), typeof(ImageSource))]
+    public class BitmapToImageSourceConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return Convert(value as Bitmap);
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotSupportedException();
+        }
+
+        public static BitmapSource Convert(Bitmap bmp)
+        {
+            if (bmp == null)
+                return null;
+            return Imaging.CreateBitmapSourceFromHBitmap(bmp.GetHbitmap(), IntPtr.Zero, Int32Rect.Empty, BitmapSizeOptions.FromEmptyOptions());
+        }
+    }
+}
