+++ Renegade X Launcher/StartupInterpreter/Interpreters/LogExp.cs
@@ -1,7 +1,7 @@
 ﻿using RxLogger;
 using System;
 
-namespace LauncherTwo.StartupInterpreter.Interpreters
+namespace LauncherTwo
 {
     public class LogExp : IStartupExpression
     {
