+++ Renegade X Launcher/Tools/InstanceHandler.cs
