+++ Renegade X Launcher/Tools/TrulyObservableCollection.cs
@@ -1,0 +1,49 @@
+﻿using System;
+using System.Collections.ObjectModel;
+using System.Collections.Specialized;
+using System.ComponentModel;
+
+namespace LauncherTwo
+{
+    // Based on http://stackoverflow.com/questions/17211462/wpf-bound-datagrid-does-not-update-items-properties
+    public class TrulyObservableCollection<T> : ObservableCollection<T>
+        where T : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler ItemPropertyChanged;
+
+        public TrulyObservableCollection()
+            : base()
+        {
+            CollectionChanged += new NotifyCollectionChangedEventHandler(TrulyObservableCollection_CollectionChanged);
+        }
+
+        void TrulyObservableCollection_CollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
+        {
+            if (e.NewItems != null)
+            {
+                foreach (Object item in e.NewItems)
+                {
+                    (item as INotifyPropertyChanged).PropertyChanged += new PropertyChangedEventHandler(item_PropertyChanged);
+                }
+            }
+            if (e.OldItems != null)
+            {
+                foreach (Object item in e.OldItems)
+                {
+                    (item as INotifyPropertyChanged).PropertyChanged -= new PropertyChangedEventHandler(item_PropertyChanged);
+                }
+            }
+        }
+
+        void item_PropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            NotifyCollectionChangedEventArgs a = new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Reset);
+            OnCollectionChanged(a);
+
+            if (ItemPropertyChanged != null)
+            {
+                ItemPropertyChanged(sender, e);
+            }
+        }
+    }
+}
