--- Renegade X Launcher/URI.cs
@@ -1,1 +1,0 @@
-﻿
